Системный дизайн/ЛР 3/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
bcrypt
python-jose
